move `LocalResourceWorkflow` to base.ts; add first batch of new tests
