+++ src/docker/workflows/confluent-local.ts
@@ -7,7 +7,7 @@ import {
   workspace,
   WorkspaceConfiguration,
 } from "vscode";
-import { findFreePort, LocalResourceContainer, LocalResourceWorkflow } from ".";
+import { findFreePort } from ".";
 import {
   ContainerCreateRequest,
   ContainerCreateResponse,
@@ -24,6 +24,7 @@ import { getLocalKafkaImageTag } from "../configs";
 import { MANAGED_CONTAINER_LABEL } from "../constants";
 import { createContainer, getContainersForImage } from "../containers";
 import { createNetwork } from "../networks";
+import { LocalResourceContainer, LocalResourceWorkflow } from "./base";
 
 const CONTAINER_NAME_PREFIX = "vscode-confluent-local-broker";
